// frontend/src/app/models/Rezervacija.ts
export class Rezervacija {
  id: number = -1;
  datum_prijave: number = 0;
  dan_posete: number = 0;
  kraj: number = 0;
  pocetak: number = 0;
  gost: string = '';
  ime_prezime: string = '';
  restoran: number = 0;
  naziv_restorana: string = '';
  adresa_restorana: string = '';
  sto: number = -1;
  broj_ljudi: number = 0;
  komentar: string = '';
  status: string = '';
  konobar: string = '';
  zahtevi: string = '';
}